pyproject.toml
[project]
name = "monetary-control"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.3",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "factory-boy>=3.3.3",
    "pytest>=9.0.2",
    "pytest-django>=4.12.0",
]
